requirements.txt
Pillow>=9.0.0
PyQt5>=5.15